requirement.txt
torch
torchvision
pillow
tk
